// src/app/vac/models/user.ts
export class User {
    userId: string;
    userFname: string;
    userLname: string;
    password: string;
    email: string;
    phone1: string;
    phone2: string;
    compCode: string;
    compName: string;
    branchCode: string;
    branchName: string;
    vatRate: string;
    version: number;
    todayDate: string;
    token: string;
    respMsg: string;    
} 